Move interfaces into an Api package.
