+++ Nancy.Rollbar/Api/IHasAccessToken.cs
@@ -1,4 +1,4 @@
-﻿namespace Nancy.Rollbar {
+﻿namespace Nancy.Rollbar.Api {
     public interface IHasAccessToken {
         string RollbarAccessToken { get; }
     }
