+++ Nancy.Rollbar/Api/IPersonFactory.cs
@@ -1,7 +1,7 @@
 ﻿using Nancy.Security;
 using Valetude.Rollbar;
 
-namespace Nancy.Rollbar {
+namespace Nancy.Rollbar.Api {
     public interface IPersonFactory{
         RollbarPerson GetPerson(IUserIdentity currentUser);
     }
