+++ Nancy.Rollbar/Api/IRollbarDataFactory.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Valetude.Rollbar;
 
-namespace Nancy.Rollbar {
+namespace Nancy.Rollbar.Api {
     public interface IRollbarDataFactory {
         RollbarData GetData(NancyContext context, Exception exception);
     }
