+++ Nancy.Rollbar/Api/IRollbarPayloadFactory.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Valetude.Rollbar;
 
-namespace Nancy.Rollbar {
+namespace Nancy.Rollbar.Api {
     public interface IRollbarPayloadFactory {
         RollbarPayload GetPayload(NancyContext context, Exception e);
     }
