+++ Nancy.Rollbar/Api/IRollbarPayloadSender.cs
@@ -1,6 +1,6 @@
 ﻿using Valetude.Rollbar;
 
-namespace Nancy.Rollbar {
+namespace Nancy.Rollbar.Api {
     public interface IRollbarPayloadSender {
         RollbarResponse SendPayload(RollbarPayload payload);
     }
